requirements.txt
openai
python-dotenv

requests